+++ client/web/src/search/results/SearchResultsInfoBar.tsx
@@ -4,17 +4,15 @@ import { mdiChevronDoubleDown, mdiChevronDoubleUp, mdiThumbUp, mdiThumbDown, mdi
 import classNames from 'classnames'
 import { useLocation } from 'react-router-dom'
 
-import { Toggle } from '@sourcegraph/branded/src/components/Toggle'
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
 import { CaseSensitivityProps, SearchPatternTypeProps } from '@sourcegraph/shared/src/search'
 import { FilterKind, findFilter } from '@sourcegraph/shared/src/search/query/query'
 import { AggregateStreamingSearchResults, StreamSearchOptions } from '@sourcegraph/shared/src/search/stream'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
-import { Button, Icon, Label, Alert, useSessionStorage, Link, Text } from '@sourcegraph/wildcard'
+import { Button, Icon, Alert, useSessionStorage, Link, Text } from '@sourcegraph/wildcard'
 
 import { AuthenticatedUser } from '../../auth'
 import { canWriteBatchChanges, NO_ACCESS_BATCH_CHANGES_WRITE, NO_ACCESS_SOURCEGRAPH_COM } from '../../batches/utils'
-import { useFeatureFlag } from '../../featureFlags/useFeatureFlag'
 import { eventLogger } from '../../tracking/eventLogger'
 
 import {
@@ -68,9 +66,6 @@ export interface SearchResultsInfoBarProps
     setSidebarCollapsed: (collapsed: boolean) => void
 
     isSourcegraphDotCom: boolean
-
-    isRankingEnabled: boolean
-    setRankingEnabled: (enabled: boolean) => void
 }
 
 /**
@@ -138,9 +133,6 @@ export const SearchResultsInfoBar: React.FunctionComponent<
         setShowMobileFilters(newShowFilters)
         props.onShowMobileFiltersChanged?.(newShowFilters)
     }
-
-    // Show/hide ranking toggle
-    const [rankingEnabled] = useFeatureFlag('search-ranking')
 
     const location = useLocation()
     const refFromCodySearch = new URLSearchParams(location.search).get('ref') === 'cody-search'
@@ -215,17 +207,6 @@ export const SearchResultsInfoBar: React.FunctionComponent<
 
                 <div className={styles.expander} />
 
-                {rankingEnabled && (
-                    <Label className={styles.toggle}>
-                        Intelligent ranking{' '}
-                        <Toggle
-                            value={props.isRankingEnabled}
-                            onToggle={() => props.setRankingEnabled(!props.isRankingEnabled)}
-                            title="Enable Ranking"
-                            className="mr-2"
-                        />
-                    </Label>
-                )}
                 <ul className="nav align-items-center">
                     <SearchActionsMenu
                         authenticatedUser={props.authenticatedUser}